LayoutTests/http/tests/referrer-policy-iframe/strict-origin/cross-origin-http-http.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of strict-origin referrer policy when cross origin.");
jsTestIsAsync = true;

if (window.internals)
    internals.setTrackingPreventionEnabled(false);

window.onmessage = function(event) {
    referrer = event.data.referrer;
    // Should be the origin, not the full URL.
    shouldBeEqualToString("referrer", "http://127.0.0.1:8000/");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}
</script>
<iframe src="http://localhost:8000/referrer-policy/resources/document.html" referrerpolicy="strict-origin"></iframe>
</body>
</html>
